Test the reception of samples from the RPi Pico emulating the PEPPER FPGA. Works for samples sent at 6 MHz SPI clock.
Samples are read dirctly into a buffer using the STM SPI blocking
read.

Since the PEPPER ADC is 24 bit and a design choice for the FPGA
was to output 32 bits per SPI word, only the first three bytes
in the receive data are valid. To keep the design simple, we
decided to capture all 4 bytes and save them instead of pruning
them.